+++ tests/pyccel/scripts/hope_benchmarks_decorators/simplify.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 
-def y(x : 'double'):
+def y(x : 'float'):
     from numpy import cos, sin
     return sin(x)**2 + (x**3 + x**2 - x - 1)/(x**2 + 2*x + 1) + cos(x)**2
